requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
openai==1.12.0
python-dotenv==1.0.1
biopython==1.83
plotly==5.19.0
matplotlib==3.8.3
altair==5.2.0
fpdf==1.7.2
tiktoken==0.6.0
requests==2.31.0
tqdm==4.66.2
scikit-learn==1.4.1
seaborn==0.13.2
nltk==3.8.1
spacy==3.7.4
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2
pytz==2024.1 